refactor: sign-up & sign-in pages into forms, buttons and tabs
